# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CrossBorder AI Demo API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/api/crowdsource/feed": {
      "get": {
        "tags": [
          "众包反馈"
        ],
        "summary": "获取近期持久化众包反馈",
        "operationId": "crowdsource_feed_api_crowdsource_feed_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 8,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrowdsourceFeedResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crowdsource/report": {
      "post": {
        "tags": [
          "众包反馈"
        ],
        "summary": "提交并持久化众包反馈",
        "operationId": "submit_crowdsource_report_api_crowdsource_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrowdsourceReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "提交成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrowdsourceSubmitResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/batch": {
      "post": {
        "tags": [
          "企业方案"
        ],
        "summary": "生成并保存企业批量方案",
        "operationId": "create_batch_plan_api_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/batch/plans": {
      "get": {
        "tags": [
          "企业方案"
        ],
        "summary": "获取近期企业批量方案",
        "operationId": "list_batch_plans_api_batch_plans_get",
        "parameters": [
          {
            "name": "company",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Company"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchHistoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/batch/csv/validate": {
      "post": {
        "tags": [
          "企业方案"
        ],
        "summary": "校验企业员工 CSV",
        "operationId": "validate_batch_csv_api_batch_csv_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCsvValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchCsvValidateResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/batch/plans/{plan_id}/export.csv": {
      "get": {
        "tags": [
          "企业方案"
        ],
        "summary": "导出企业方案 CSV",
        "operationId": "export_batch_plan_api_batch_plans__plan_id__export_csv_get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BatchCsvError": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "row",
          "message"
        ],
        "title": "BatchCsvError"
      },
      "BatchCsvValidateRequest": {
        "properties": {
          "csv_text": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Csv Text"
          }
        },
        "type": "object",
        "required": [
          "csv_text"
        ],
        "title": "BatchCsvValidateRequest"
      },
      "BatchCsvValidateResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "employees": {
            "items": {
              "$ref": "#/components/schemas/BatchEmployee"
            },
            "type": "array",
            "title": "Employees"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/BatchCsvError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "employees",
          "errors"
        ],
        "title": "BatchCsvValidateResponse"
      },
      "BatchEmployee": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 60,
            "title": "Name",
            "default": ""
          },
          "origin_id": {
            "type": "string",
            "minLength": 1,
            "title": "Origin Id"
          },
          "destination_id": {
            "type": "string",
            "minLength": 1,
            "title": "Destination Id"
          },
          "arrival_deadline": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}$",
            "title": "Arrival Deadline"
          },
          "preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictionPreferences"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "origin_id",
          "destination_id",
          "arrival_deadline"
        ],
        "title": "BatchEmployee"
      },
      "BatchHistoryRecord": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "request": {
            "additionalProperties": true,
            "type": "object",
            "title": "Request"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "company",
          "date",
          "request",
          "result",
          "created_at"
        ],
        "title": "BatchHistoryRecord"
      },
      "BatchHistoryResponse": {
        "properties": {
          "plans": {
            "items": {
              "$ref": "#/components/schemas/BatchHistoryRecord"
            },
            "type": "array",
            "title": "Plans"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "plans",
          "total"
        ],
        "title": "BatchHistoryResponse"
      },
      "BatchPlanItem": {
        "properties": {
          "employee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Employee Id"
          },
          "recommended_port": {
            "type": "string",
            "title": "Recommended Port"
          },
          "departure_time": {
            "type": "string",
            "title": "Departure Time"
          },
          "total_time": {
            "type": "integer",
            "title": "Total Time"
          },
          "late_risk_percent": {
            "type": "integer",
            "title": "Late Risk Percent"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "max_budget": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Budget"
          },
          "within_budget": {
            "type": "boolean",
            "title": "Within Budget"
          }
        },
        "type": "object",
        "required": [
          "employee_id",
          "recommended_port",
          "departure_time",
          "total_time",
          "late_risk_percent",
          "priority",
          "max_budget",
          "within_budget"
        ],
        "title": "BatchPlanItem"
      },
      "BatchRequest": {
        "properties": {
          "company": {
            "type": "string",
            "minLength": 1,
            "title": "Company"
          },
          "employees": {
            "items": {
              "$ref": "#/components/schemas/BatchEmployee"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Employees"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "preferences": {
            "$ref": "#/components/schemas/PredictionPreferences"
          }
        },
        "type": "object",
        "required": [
          "company",
          "employees",
          "date"
        ],
        "title": "BatchRequest"
      },
      "BatchResponse": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "plan": {
            "items": {
              "$ref": "#/components/schemas/BatchPlanItem"
            },
            "type": "array",
            "title": "Plan"
          },
          "summary": {
            "$ref": "#/components/schemas/BatchSummary"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "company",
          "date",
          "plan",
          "summary"
        ],
        "title": "BatchResponse"
      },
      "BatchSummary": {
        "properties": {
          "employee_count": {
            "type": "integer",
            "title": "Employee Count"
          },
          "avg_commute_time": {
            "type": "integer",
            "title": "Avg Commute Time"
          },
          "high_risk_count": {
            "type": "integer",
            "title": "High Risk Count"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          }
        },
        "type": "object",
        "required": [
          "employee_count",
          "avg_commute_time",
          "high_risk_count",
          "recommendation"
        ],
        "title": "BatchSummary"
      },
      "CrossingChannel": {
        "type": "string",
        "enum": [
          "traveller",
          "vehicle",
          "cargo"
        ],
        "title": "CrossingChannel"
      },
      "CrowdLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "CrowdLevel"
      },
      "CrowdsourceFeedResponse": {
        "properties": {
          "reports": {
            "items": {
              "$ref": "#/components/schemas/CrowdsourceRecord"
            },
            "type": "array",
            "title": "Reports"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "reports",
          "total"
        ],
        "title": "CrowdsourceFeedResponse"
      },
      "CrowdsourceRecord": {
        "properties": {
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id",
            "default": "demo-user"
          },
          "port": {
            "type": "string",
            "minLength": 1,
            "title": "Port"
          },
          "actual_wait_time": {
            "type": "integer",
            "maximum": 180.0,
            "minimum": 0.0,
            "title": "Actual Wait Time"
          },
          "crowd_level": {
            "$ref": "#/components/schemas/CrowdLevel"
          },
          "comment": {
            "type": "string",
            "maxLength": 160,
            "title": "Comment",
            "default": ""
          },
          "forecast_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Run Id"
          },
          "forecast_port_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Port Id"
          },
          "direction": {
            "$ref": "#/components/schemas/TravelDirection",
            "default": "hong_kong_to_shenzhen"
          },
          "channel": {
            "$ref": "#/components/schemas/CrossingChannel",
            "default": "traveller"
          },
          "is_real_observation": {
            "type": "boolean",
            "title": "Is Real Observation",
            "default": false
          },
          "training_consent": {
            "type": "boolean",
            "title": "Training Consent",
            "default": false
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "time_label": {
            "type": "string",
            "title": "Time Label"
          },
          "quality_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Quality Score"
          },
          "quality_level": {
            "$ref": "#/components/schemas/ReportQualityLevel"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "used_for_prediction": {
            "type": "boolean",
            "title": "Used For Prediction"
          },
          "source_type": {
            "$ref": "#/components/schemas/ObservationSource"
          },
          "wait_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait Started At"
          },
          "wait_ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait Ended At"
          },
          "eligible_for_v2_label": {
            "type": "boolean",
            "title": "Eligible For V2 Label"
          }
        },
        "type": "object",
        "required": [
          "port",
          "actual_wait_time",
          "crowd_level",
          "id",
          "timestamp",
          "time_label",
          "quality_score",
          "quality_level",
          "expires_at",
          "used_for_prediction",
          "source_type",
          "eligible_for_v2_label"
        ],
        "title": "CrowdsourceRecord"
      },
      "CrowdsourceReport": {
        "properties": {
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id",
            "default": "demo-user"
          },
          "port": {
            "type": "string",
            "minLength": 1,
            "title": "Port"
          },
          "actual_wait_time": {
            "type": "integer",
            "maximum": 180.0,
            "minimum": 0.0,
            "title": "Actual Wait Time"
          },
          "crowd_level": {
            "$ref": "#/components/schemas/CrowdLevel"
          },
          "comment": {
            "type": "string",
            "maxLength": 160,
            "title": "Comment",
            "default": ""
          },
          "forecast_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Run Id"
          },
          "forecast_port_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Port Id"
          },
          "direction": {
            "$ref": "#/components/schemas/TravelDirection",
            "default": "hong_kong_to_shenzhen"
          },
          "channel": {
            "$ref": "#/components/schemas/CrossingChannel",
            "default": "traveller"
          },
          "is_real_observation": {
            "type": "boolean",
            "title": "Is Real Observation",
            "default": false
          },
          "training_consent": {
            "type": "boolean",
            "title": "Training Consent",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "port",
          "actual_wait_time",
          "crowd_level"
        ],
        "title": "CrowdsourceReport"
      },
      "CrowdsourceSubmitResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "points_earned": {
            "type": "integer",
            "title": "Points Earned"
          },
          "model_updated": {
            "type": "boolean",
            "title": "Model Updated"
          },
          "report": {
            "$ref": "#/components/schemas/CrowdsourceRecord"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "forecast_feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForecastFeedbackLink"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success",
          "points_earned",
          "model_updated",
          "report",
          "message"
        ],
        "title": "CrowdsourceSubmitResponse"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "ForecastFeedbackLink": {
        "properties": {
          "forecast_run_id": {
            "type": "string",
            "title": "Forecast Run Id"
          },
          "forecast_port_id": {
            "type": "string",
            "title": "Forecast Port Id"
          },
          "linked": {
            "type": "boolean",
            "title": "Linked"
          },
          "labeled": {
            "type": "boolean",
            "title": "Labeled"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "forecast_run_id",
          "forecast_port_id",
          "linked",
          "labeled"
        ],
        "title": "ForecastFeedbackLink"
      },
...
      "ObservationSource": {
        "type": "string",
        "enum": [
          "demo_seed",
          "demo_entry",
          "crowdsource_observation",
          "partner",
          "official"
        ],
        "title": "ObservationSource"
      },
...
      "PredictionPreferences": {
        "properties": {
          "priority": {
            "$ref": "#/components/schemas/Priority",
            "default": "balanced"
          },
          "max_budget": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Budget"
          }
        },
        "type": "object",
        "title": "PredictionPreferences"
      },
...
      "Priority": {
        "type": "string",
        "enum": [
          "fastest",
          "cheapest",
          "balanced"
        ],
        "title": "Priority"
      },
...
      "ReportQualityLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "ReportQualityLevel"
      },
...
      "TravelDirection": {
        "type": "string",
        "enum": [
          "hong_kong_to_shenzhen",
          "shenzhen_to_hong_kong"
        ],
        "title": "TravelDirection"
      },
...
    }
  }
}